clipboard-apis/integration_with_execcommand_can_prevent_default_action_cut_on_ifrcontentwindow.html

<!DOCTYPE html>
<html><head>
    <meta charset="UTF-8">
    <title>Clipboard API - integration with execCommand, can prevent default action  - cut on 'ifr.contentWindow'</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/_lib.js"></script>
    <script>
/** integration with execCommand, can prevent default action  */
/* Events: copy cut */
/* paste data: "clipboard text" */
/* Test HTML: <iframe id="ifr"></iframe>*/
/* Targets: 'ifr.contentWindow' */
/* External pass condition - clipboard data: "clipboard text" */

function clipboard_api_test(e){
    e.preventDefault();
}

function triggerTestManually(){
  document.getElementsByTagName('p')[0].firstChild.data='FAILED'; // if event doesn't fire, this test failed..
  var ifr=document.getElementById('ifr');
  var doc=ifr.contentDocument;
  doc.designMode='on';
  doc.body.innerHTML='<p>Frame contents</p>';
  ifr.contentWindow.focus();
  var range=doc.createRange();
  range.selectNodeContents(doc.getElementsByTagName('p')[0]);
  doc.defaultView.getSelection().removeAllRanges();
  doc.defaultView.getSelection().addRange(range);
  var returnValue = doc.execCommand(window.eventType, false, false);
  /* Main pass condition here is clipboard contents - but we can add a couple of assertations here */
  assert_false(returnValue, 'execCommand() returns false when event default action is prevented');
  /* This assertation is especially relevant for 'cut' */
  assert_equals(doc.getElementsByTagName('p')[0].firstChild.textContent, 'Frame contents');
}

    </script>
</head>
<body>

    <p>FAILED (This TC requires JavaScript enabled)</p>
    <iframe id="ifr"></iframe>
    <div id="log"></div>
    <script>
    var dataToPaste="clipboard text" ;
    var extPassCond="clipboard text" ;
    var eventTarget='ifr.contentWindow';
    var eventType='cut';
    setup(function(){
            setupTest( eventTarget, eventType, dataToPaste, extPassCond );
        });
    </script>
</body></html>
